pyproject.toml
[tool.poetry]
name = "package-name"
version = "0.1.0"
description = ""
authors = ["Alina Davydenko <Alinadavydenko620@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
aiogram = "^3.13.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.2"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
